LayoutTests/platform/blackberry/fast/dom/Range/getBoundingClientRect-getClientRects-relative-to-viewport-expected.txt

PASS unScrolledBoundingBox.top - scrolledDownBoundingBox.top is 50
PASS unScrolledBoundingBoxes[0].top - scrolledDownBoundingBoxes[0].top is 50
FAIL unScrolledBoundingBox.left - scrolledRightBoundingBox.left should be 50. Was 0.
FAIL unScrolledBoundingBoxes[0].left - scrolledRightBoundingBoxes[0].left should be 50. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
